smoke: csp
